# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/work/create": {
            "post": {
                "tags": [
                    "Works"
                ],
                "summary": "Add works to employee",
                "operationId": "7403e9994d83b24d56bd59f1b66aa8b5",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "employee_id": {
                                        "description": "User ID where type is employee",
                                        "type": "integer"
                                    },
                                    "images[0][image]": {
                                        "description": "Image work 1",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "images[1][image]": {
                                        "description": "Image work 2",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "images[2][image]": {
                                        "description": "Image work 3",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "images[3][image]": {
                                        "description": "Image work 4",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "images[4][image]": {
                                        "description": "Image work 5",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "videos[0][video]": {
                                        "description": "Video work 1",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "videos[1][video]": {
                                        "description": "Video work 2",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "videos[2][video]": {
                                        "description": "Video work 3",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "videos[3][video]": {
                                        "description": "Video work 4",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "videos[4][video]": {
                                        "description": "Video work 5",
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee works added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "successfully"
                                        },
                                        "employee": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}